django_wellness/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
axios==0.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
openai==1.40.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
rich==13.7.1
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
